Chaco

+++ client/src/components/cart/Cart.jsx
@@ -63,7 +63,7 @@ function Cart({products, getAllCart, getcart, onlineUser, updateCart, completeCa
   const shipping = 400; // envío
   const taxes = useRef(0) // impuesto
   const total = useRef(0) // total
-  const total = useRef(0) // total
+
   
   const handleCantidadDelProducto = (id, price) => {
 
@@ -201,17 +201,56 @@ function Cart({products, getAllCart, getcart, onlineUser, updateCart, completeCa
 
     if(typeof onlineUser == "object"){
     var order = getcart[0].order_id;
-    deleteProductCart(order, id)
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You are about to remove the product from the cart",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+        //console.log(result);
+      if (result.value) {
+        deleteProductCart(order, id)
+        Swal.fire(
+          'Deleted!',
+          'The product has been clear.',
+          'success'
+        );
+      }
+    })
+
+   
   }else{
     let arry = [];
     ls.get('idProducts').forEach(function(ele){
         return arry.push(parseInt(ele))
       });
-    const result = arry.filter(word => word !== id);
-      console.log(result);
-      ls.set('idProducts', []);
-      ls.set('idProducts', result);
-      history.push('/cart')
+    const resultadoaa = arry.filter(word => word !== id);
+      //console.log(result);
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You are about to remove the product from the cart",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+          //console.log(result);
+        if (result.value) {
+          ls.set('idProducts', []);
+          ls.set('idProducts', resultadoaa);
+          history.push('/cart')
+          Swal.fire(
+            'Deleted!',
+            'The product has been clear.',
+            'success'
+          );
+        }
+      })
+      
 
 
 
@@ -278,7 +317,7 @@ function Cart({products, getAllCart, getcart, onlineUser, updateCart, completeCa
 									<input class="cart_quantity_input" type="quantity" name="quantity" min="1" max={e.stock} value="1" autocomplete="off" size="2"/>
 									<a class="cart_quantity_down" id="boton_cantidad_menos"> + </a> */}
 
-                  <input min="1" max={e.stock} defaultValue="1" type="number" id={e.id+"input"} onClick={() => handleCantidadDelProducto(e.id, e.price)} className="form-control inputCantidad" style={{width: "60%", borderColor:"orange"}} />
+                  <input min="1" max={e.stock} defaultValue="1" type="number" id={e.id+"input"} onClick={() => handleCantidadDelProducto(e.id, e.price)} className="form-control inputCantidad" style={{width: "90%", borderColor:"orange", textAlign:'center'}} />
 
 								</div>
 							</td>
